refactor(orchestrator): move pure emission helpers to event_emitter module
